requirements.txt
asgiref==3.5.0
Django==4.0.4
gunicorn==20.1.0
mysqlclient==2.1.0
python-dotenv==0.20.0
sqlparse==0.4.2